[Misc] extract out helper into validator_component.h
and flatten the structs
